reorganize search tags/values
